requirements.txt
tensorflow==2.15.0
fastapi
uvicorn
jinja2
pydantic
pandas
numpy
scikit-learn>=1.3
python-dotenv
tweepy>=4.14
dvc
pytest
flake8
pywin32==311; sys_platform == 'win32'
matplotlib>=3.7
scipy>=1.10
mlflow>=2.10
torch>=2.0
transformers>=4.38
datasets
requests
pyyaml
